Disable test on firefox due not supporting selecting empty files

+++ src/content/datachannel/filetransfer/js/test.js
@@ -20,9 +20,10 @@ function sendFile(t, path) {
   .then(function() {
     // Wait for the received element to be displayed.
     return driver.wait(webdriver.until.elementIsVisible(
-        driver.findElement(webdriver.By.id('received'))));
+        driver.findElement(webdriver.By.id('received'))), 90 * 1000);
   })
   .then(function() {
+    t.pass('received element found');
     t.end();
   })
   .then(null, function(err) {
@@ -49,14 +50,19 @@ test('Filetransfer via Datachannels: video', function(t) {
 });
 
 test('Filetransfer via Datachannels: empty file', function(t) {
-  var fs = require('fs');
-  var emptyFilePath = '/src/content/datachannel/filetransfer/emptyFile';
-  fs.writeFile(process.cwd() + emptyFilePath, '');
+  // TODO: Remove when supported in Firefox.
+  if (process.env.BROWSER === 'firefox') {
+    t.skip('Empty file selection is not supported on firefox');
+    t.end();
+  } else {
+    var fs = require('fs');
+    var emptyFilePath = '/src/content/datachannel/filetransfer/emptyFile';
+    // Create empty file.
+    fs.writeFileSync(process.cwd() + emptyFilePath, '');
 
-  sendFile(t, process.cwd() + emptyFilePath);
+    sendFile(t, process.cwd() + emptyFilePath);
 
-  // Remove the empty file
-  fs.unlink(process.cwd() + emptyFilePath);
+    // Remove the empty file.
+    fs.unlink(process.cwd() + emptyFilePath);
+  }
 });
-
-
